スリップ距離補正機能の強化とエンコーダ補正の追加 - LOG_NUM_32BITを4から5に変更 - スリップ距離補正パラメータの調整 - calcXYcie関数の引数を変更 - 補正後の現在速度を管理するencCurrentCorr_pを追加 - ログ出力にencCurrentCorr_pを追加

+++ robotrace_v2/Core/Inc/courseAnalysis.h
@@ -88,7 +88,7 @@ float asignVelocity(int16_t ROC);
 int cmpfloat(const void *n1, const void *n2);
 int16_t readLogTest(int logNumber);
 int16_t calcXYcies(int logNumber);
-void calcXYcie(float encpulse, float angVelo, float dt);
+void calcXYcie(int16_t encpulse, float angVelo, float dt);
 void clearXYcie(void);
 void setShortCutTarget(void);
 void processMarkerEvent(void);
